Added information about Pack archives
